Move definition of main() to Main.cpp
